requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==2.2
django-filter==2.3.0
django-rosetta==0.9.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
importlib-metadata==1.7.0
kavenegar==1.1.2
kombu==4.6.11
Markdown==3.2.2
Pillow==7.2.0
polib==1.1.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-memcached==1.59
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
